Extract dispatcher with map to helper
