+++ src/androidTest/java/helper/IntegrationHelper.java
@@ -9,11 +9,17 @@ import com.google.common.reflect.TypeToken;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 
+import java.io.IOException;
 import java.lang.reflect.Constructor;
 import java.lang.reflect.Type;
+import java.net.URI;
 import java.util.Collections;
 import java.util.List;
+import java.util.Map;
 
+import fake.HttpClientMock;
+import fake.HttpResponseMock;
+import fake.HttpResponseMockDispatcher;
 import io.split.android.client.SplitClientConfig;
 import io.split.android.client.SplitFactory;
 import io.split.android.client.SplitFactoryImpl;
@@ -21,9 +27,11 @@ import io.split.android.client.api.Key;
 import io.split.android.client.dtos.Event;
 import io.split.android.client.dtos.TestImpressions;
 import io.split.android.client.network.HttpClient;
+import io.split.android.client.network.HttpMethod;
 import io.split.android.client.service.synchronizer.SynchronizerSpy;
 import io.split.android.client.storage.db.SplitRoomDatabase;
 import io.split.android.client.utils.Logger;
+import io.split.sharedtest.fake.HttpStreamResponseMock;
 
 public class IntegrationHelper {
     public static final int NEVER_REFRESH_RATE = 999999;
@@ -168,5 +176,68 @@ public class IntegrationHelper {
 
     public static String streamingEnabledV1Token() {
         return "{\"connDelay\":0,\"pushEnabled\":true,\"token\":\"eyJhbGciOiJIUzI1NiIsImtpZCI6IjVZOU05US5pSGZUUmciLCJ0eXAiOiJKV1QifQ.eyJ4LWFibHktY2FwYWJpbGl0eSI6IntcIk56TTJNREk1TXpjMF9NVGd5TlRnMU1UZ3dOZz09X01Ua3pOamd3TURFNE1BPT1fbXlTZWdtZW50c1wiOltcInN1YnNjcmliZVwiXSxcIk56TTJNREk1TXpjMF9NVGd5TlRnMU1UZ3dOZz09X01qWXhNRE0yTkRjd09RPT1fbXlTZWdtZW50c1wiOltcInN1YnNjcmliZVwiXSxcIk56TTJNREk1TXpjMF9NVGd5TlRnMU1UZ3dOZz09X3NwbGl0c1wiOltcInN1YnNjcmliZVwiXSxcImNvbnRyb2xfcHJpXCI6W1wic3Vic2NyaWJlXCIsXCJjaGFubmVsLW1ldGFkYXRhOnB1Ymxpc2hlcnNcIl0sXCJjb250cm9sX3NlY1wiOltcInN1YnNjcmliZVwiLFwiY2hhbm5lbC1tZXRhZGF0YTpwdWJsaXNoZXJzXCJdfSIsIngtYWJseS1jbGllbnRJZCI6ImNsaWVudElkIiwiZXhwIjoxNjQ4NjU2MjU4LCJpYXQiOjE2NDg2NTI2NTh9.MWwudv3kafKr-gVeqt-ClLAkCngZsDhdWx-dwqM9rxs\"}";
+    }
+
+    /**
+     * Builds a dispatcher with the given responses.
+     *
+     * @param responses The responses to be returned by the dispatcher. The keys are url paths.
+     * @return The dispatcher to be used in {@link HttpClientMock}
+     */
+    public static HttpResponseMockDispatcher buildDispatcher(Map<String, ResponseClosure> responses) {
+        return buildDispatcher(responses, Collections.emptyMap());
+    }
+
+    /**
+     * Builds a dispatcher with the given responses.
+     *
+     * @param responses The responses to be returned by the dispatcher. The keys are url paths.
+     * @param streamingResponses The streaming responses to be returned by the dispatcher. The keys are url paths.
+     * @return The dispatcher to be used in {@link HttpClientMock}
+     */
+    public static HttpResponseMockDispatcher buildDispatcher(Map<String, ResponseClosure> responses, Map<String, StreamingResponseClosure> streamingResponses) {
+        return new HttpResponseMockDispatcher() {
+            @Override
+            public HttpResponseMock getResponse(URI uri, HttpMethod method, String body) {
+                String path = uri.getPath().replace("/api/", "");
+                if (responses.containsKey(path)) {
+                    return responses.get(path).onResponse(uri, method, body);
+                } else {
+                    return new HttpResponseMock(200, "{}");
+                }
+            }
+
+            @Override
+            public HttpStreamResponseMock getStreamResponse(URI uri) {
+                try {
+                    String path = uri.getPath().replace("/api/", "");
+                    if (streamingResponses.containsKey(path)) {
+                        return streamingResponses.get(path).onResponse(uri);
+                    } else {
+                        return new HttpStreamResponseMock(200, null);
+                    }
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
+
+                return null;
+            }
+        };
+    }
+
+    /**
+     * A simple interface to allow us to define the response for a given path
+     */
+    public interface ResponseClosure {
+        HttpResponseMock onResponse(URI uri,
+                                    HttpMethod httpMethod,
+                                    String body);
+    }
+
+    /**
+     * A simple interface to allow us to define the streaming response for a given path
+     */
+    public interface StreamingResponseClosure {
+        HttpStreamResponseMock onResponse(URI uri);
     }
 }
